+++ NEWS.md
@@ -1,6 +1,7 @@
 # stevethemes 0.2.0
 
 - Add `are_fonts_installed()` for further user diagnostics.
+- Add `how_to_install_fonts()` to assist the user in how to install fonts on their operating system.
 - Add support for Merriweather font, which is a close cousin to the Palatino family.
 - Check for installation of Merriweather upon startup.